+++ CHANGELOG.md
@@ -5,7 +5,10 @@ PowerModels.jl Change Log
 - Droped lower bounds on voltage variables
 - Added all error messages to memento log
 - Fixed minor typos in docs
-- Added a corner tolerance fix for OBBT 
+- Added a corner tolerance fix for OBBT
+- Added ability to parse default values in PTI files
+- Fixed bug in parsing of PTI transformers (#459)
+- Refactor PTI parsing function names with `_` to indicate internal use only
 
 ### v0.9.3
 - Added support for heterogeneous cost functions